Cleaning code
Removing unused branches and code duplication
